requirements.txt
Pillow
click
python-vlc
rich